+++ src/main/java/org/usf/trace/api/server/CacheController.java
@@ -1,0 +1,28 @@
+package org.usf.trace.api.server;
+
+import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
+
+import java.util.Collection;
+
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import org.usf.traceapi.core.Session;
+
+import lombok.RequiredArgsConstructor;
+
+@CrossOrigin
+@RestController
+@RequestMapping(value = "cache", produces = APPLICATION_JSON_VALUE)
+@RequiredArgsConstructor
+public class CacheController {
+	
+    private final SessionQueueService queue;
+
+    @GetMapping
+    public Collection<Session> getCache(){
+    	return queue.waitList();
+    }
+    
+}
